added new clientUtils for getting EMR Type
