+++ lib/cjs/Client/ClientFactory.js
@@ -62,13 +62,14 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.LAUNCH = void 0;
+exports.instanceOfJWT = exports.LAUNCH = void 0;
 var fhirclient_1 = __importStar(require("fhirclient")), FHIR = fhirclient_1;
+var __1 = require("..");
 var SmartLaunchHandler_1 = require("../Launcher/SmartLaunchHandler");
 var CernerClient_1 = __importDefault(require("./CernerClient"));
+var ECWClient_1 = __importDefault(require("./ECWClient"));
 var EpicClient_1 = __importDefault(require("./EpicClient"));
 var SmartHealthClient_1 = __importDefault(require("./SmartHealthClient"));
-var ECWClient_1 = __importDefault(require("./ECWClient"));
 var LAUNCH;
 (function (LAUNCH) {
     LAUNCH[LAUNCH["EMR"] = 0] = "EMR";
@@ -83,43 +84,13 @@ var LAUNCH;
 function instanceOfJWT(object) {
     return object.client_id !== undefined;
 }
+exports.instanceOfJWT = instanceOfJWT;
 /**
 Represents the ClientFactory class for creating EMR clients.
 */
 var ClientFactory = /** @class */ (function () {
     function ClientFactory() {
     }
-    /**
-     * The function `getEMRType` determines the type of Electronic Medical Record (EMR) based on the provided client or token.
-     * @param {SubClient | JWT} client - The parameter `clientOrToken` can be either a `SubClient` object or a JWT (JSON Web Token).
-     * @returns the type of Electronic Medical Record (EMR) based on the input parameter. The possible return values are EMR.CERNER, EMR.SMART, EMR.EPIC, or EMR.NONE.
-    */
-    ClientFactory.prototype.getEMRType = function (clientOrToken) {
-        function isClient(input) {
-            return input.state.serverUrl !== undefined;
-        }
-        var EMR_MAPPING = [
-            { substring: "cerner", emr: SmartLaunchHandler_1.EMR.CERNER },
-            { substring: "smarthealthit", emr: SmartLaunchHandler_1.EMR.SMART },
-            { substring: "epic", emr: SmartLaunchHandler_1.EMR.EPIC },
-            { substring: "ecw", emr: SmartLaunchHandler_1.EMR.ECW }
-        ];
-        if (isClient(clientOrToken)) {
-            var serverUrl = clientOrToken.state.serverUrl;
-            for (var _i = 0, EMR_MAPPING_1 = EMR_MAPPING; _i < EMR_MAPPING_1.length; _i++) {
-                var _a = EMR_MAPPING_1[_i], substring = _a.substring, emr = _a.emr;
-                if (serverUrl.includes(substring)) {
-                    return emr;
-                }
-            }
-        }
-        else {
-            if ("epic.eci" in clientOrToken) {
-                return SmartLaunchHandler_1.EMR.EPIC;
-            }
-        }
-        return SmartLaunchHandler_1.EMR.NONE;
-    };
     /**
      * The function `createEMRClient` creates an EMR client based on the specified launch type.
      * @param {LAUNCH} launchType - The `launchType` parameter is an optional parameter of type `LAUNCH` that specifies the type of EMR launch. It has a default value
@@ -165,7 +136,7 @@ var ClientFactory = /** @class */ (function () {
         return __awaiter(this, void 0, void 0, function () {
             var emrType;
             return __generator(this, function (_a) {
-                emrType = this.getEMRType(fhirClient);
+                emrType = __1.ClientUtils.getEMRType(fhirClient);
                 switch (emrType) {
                     case SmartLaunchHandler_1.EMR.EPIC:
                         return [2 /*return*/, new EpicClient_1.default(fhirClient)];
@@ -204,7 +175,7 @@ var ClientFactory = /** @class */ (function () {
         });
     };
     ClientFactory.prototype.getEmrEndpoints = function (object) {
-        switch (this.getEmrTypeFromObject(object)) {
+        switch (__1.ClientUtils.getEmrTypeFromObject(object)) {
             case SmartLaunchHandler_1.EMR.EPIC:
                 return EpicClient_1.default.getEndpoints();
             case SmartLaunchHandler_1.EMR.CERNER:
@@ -214,20 +185,6 @@ var ClientFactory = /** @class */ (function () {
             default:
                 throw new Error('EMR type not defined.');
         }
-    };
-    /**
-     * The function `getEmrTypeFromObject` takes an object as input and returns the corresponding EMR type if the object is of type JWT or EMR, otherwise it throws an
-     * error.
-     * @param {unknown} object - The `object` parameter is of type `unknown`, which means it can be any type. It is used as input to determine the EMR (Electronic
-     * Medical Record) type. The function checks if the `object` is an instance of JWT (JSON Web Token) or EMR, and returns
-     * @returns an EMR (Electronic Medical Record) object.
-     */
-    ClientFactory.prototype.getEmrTypeFromObject = function (object) {
-        if (instanceOfJWT(object))
-            return this.getEMRType(object);
-        if ((0, SmartLaunchHandler_1.instanceOfEmr)(object))
-            return object;
-        throw new Error('Invalid object type.');
     };
     return ClientFactory;
 }());
