+++ lib/esm/Client/ClientFactory.js
@@ -8,11 +8,12 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 import smart, * as FHIR from "fhirclient";
-import { EMR, instanceOfEmr } from "../Launcher/SmartLaunchHandler";
+import { ClientUtils } from "..";
+import { EMR } from "../Launcher/SmartLaunchHandler";
 import CernerClient from "./CernerClient";
+import ECWClient from "./ECWClient";
 import EpicClient from "./EpicClient";
 import SmartHealthClient from "./SmartHealthClient";
-import ECWClient from "./ECWClient";
 export var LAUNCH;
 (function (LAUNCH) {
     LAUNCH[LAUNCH["EMR"] = 0] = "EMR";
@@ -24,43 +25,13 @@ export var LAUNCH;
  * @param {unknown} object - The `object` parameter is of type `unknown`, which means it can be any type.
  * @returns a boolean value.
  */
-function instanceOfJWT(object) {
+export function instanceOfJWT(object) {
     return object.client_id !== undefined;
 }
 /**
 Represents the ClientFactory class for creating EMR clients.
 */
 export default class ClientFactory {
-    /**
-     * The function `getEMRType` determines the type of Electronic Medical Record (EMR) based on the provided client or token.
-     * @param {SubClient | JWT} client - The parameter `clientOrToken` can be either a `SubClient` object or a JWT (JSON Web Token).
-     * @returns the type of Electronic Medical Record (EMR) based on the input parameter. The possible return values are EMR.CERNER, EMR.SMART, EMR.EPIC, or EMR.NONE.
-    */
-    getEMRType(clientOrToken) {
-        function isClient(input) {
-            return input.state.serverUrl !== undefined;
-        }
-        const EMR_MAPPING = [
-            { substring: "cerner", emr: EMR.CERNER },
-            { substring: "smarthealthit", emr: EMR.SMART },
-            { substring: "epic", emr: EMR.EPIC },
-            { substring: "ecw", emr: EMR.ECW }
-        ];
-        if (isClient(clientOrToken)) {
-            const serverUrl = clientOrToken.state.serverUrl;
-            for (const { substring, emr } of EMR_MAPPING) {
-                if (serverUrl.includes(substring)) {
-                    return emr;
-                }
-            }
-        }
-        else {
-            if ("epic.eci" in clientOrToken) {
-                return EMR.EPIC;
-            }
-        }
-        return EMR.NONE;
-    }
     /**
      * The function `createEMRClient` creates an EMR client based on the specified launch type.
      * @param {LAUNCH} launchType - The `launchType` parameter is an optional parameter of type `LAUNCH` that specifies the type of EMR launch. It has a default value
@@ -93,7 +64,7 @@ export default class ClientFactory {
     }
     createSmarterFhirClient(fhirClient) {
         return __awaiter(this, void 0, void 0, function* () {
-            const emrType = this.getEMRType(fhirClient);
+            const emrType = ClientUtils.getEMRType(fhirClient);
             switch (emrType) {
                 case EMR.EPIC:
                     return new EpicClient(fhirClient);
@@ -127,7 +98,7 @@ export default class ClientFactory {
         });
     }
     getEmrEndpoints(object) {
-        switch (this.getEmrTypeFromObject(object)) {
+        switch (ClientUtils.getEmrTypeFromObject(object)) {
             case EMR.EPIC:
                 return EpicClient.getEndpoints();
             case EMR.CERNER:
@@ -137,20 +108,6 @@ export default class ClientFactory {
             default:
                 throw new Error('EMR type not defined.');
         }
-    }
-    /**
-     * The function `getEmrTypeFromObject` takes an object as input and returns the corresponding EMR type if the object is of type JWT or EMR, otherwise it throws an
-     * error.
-     * @param {unknown} object - The `object` parameter is of type `unknown`, which means it can be any type. It is used as input to determine the EMR (Electronic
-     * Medical Record) type. The function checks if the `object` is an instance of JWT (JSON Web Token) or EMR, and returns
-     * @returns an EMR (Electronic Medical Record) object.
-     */
-    getEmrTypeFromObject(object) {
-        if (instanceOfJWT(object))
-            return this.getEMRType(object);
-        if (instanceOfEmr(object))
-            return object;
-        throw new Error('Invalid object type.');
     }
 }
 //# sourceMappingURL=ClientFactory.js.map
